requirements.txt
sqlalchemy>=1.4,<2.0
argon2_cffi>=20.1.0
beautifulsoup4>=4.11.1
celery[redis]>=5.2.7
cryptography>=41.0.0
email-validator>=1.1.3
emoji>=2.0.0
feedgen>=0.9.0
ffmpeg-python>=0.2.0
flask_sqlalchemy>=2.5.1
Flask-Caching>=1.10.1
Flask-CeleryExt>=0.4.1
flask-crontab>=0.1.2
Flask-Limiter>=3.3.0
Flask-Login>=0.6.0
flask-marshmallow>=0.14.0
Flask-Migrate>=4.0.0
Flask-RESTful>=0.3.9
Flask-WTF>=1.1.0
Flask>=2.3.0
gallery-dl>=1.25.0
gunicorn>=21.0.0
humanize>=4.0.0
lxml>=4.9.0
markdown2>=2.4.3
marshmallow-dataclass[enum,union]>=8.5.8
marshmallow-sqlalchemy>=0.28.0
marshmallow>=3.14.1
mdx_bleach>=0.1.4
passlib>=1.7.4
pika>=1.3.0
Pillow>=9.5.0
psycopg2-binary>=2.9.5
pybase64>=1.2.0
python-dateutil>=2.8.2
redis>=4.5.0
regex>=2022.10.31
requests>=2.28.0
sqlalchemy-utils>=0.40.0
traceback-with-variables>=2.0.4
ujson>=5.5.0
url-normalize>=1.4.3
Werkzeug>=2.3.0
xmltodict>=0.13.0
Zenora>=0.0.3
flask-qrcode==3.1.0
pyotp==2.8.0
# Dev only - flask-debugtoolbar is loaded conditionally in debug mode
flask-debugtoolbar==0.13